fixed fmt issue

+++ src/FundMe.sol
@@ -4,7 +4,7 @@ pragma solidity >=0.8.7 <0.9.0;
 
 import {Test, console} from "forge-std/Test.sol";
 import {PriceConverter} from "./PriceConverter.sol";
-import {AggregatorV3Interface} from "@chainlink/contracts/src/v0.8/interfaces/AggregatorV3Interface.sol";
+import {AggregatorV3Interface} from "@chainlink/contracts/src/v0.8/shared/interfaces/AggregatorV3Interface.sol";
 
 error FundMe__NotOwner();
 
@@ -24,6 +24,7 @@ contract FundMe {
         uint256 amount;
         uint256 time;
     }
+
     Transaction[] private s_transactions;
     mapping(address => uint256) private s_addressToAmountFunded;
     address[] private s_funders;
@@ -34,10 +35,7 @@ contract FundMe {
         s_bank = 0;
         s_minimumUSD = 50;
         priceFeed = AggregatorV3Interface(_priceFeed);
-        console.log(
-            "FundMe's balance in constructor: %s",
-            address(this).balance
-        );
+        console.log("FundMe's balance in constructor: %s", address(this).balance);
     }
 
     modifier onlyOwner() {
@@ -54,51 +52,35 @@ contract FundMe {
             msg.value * getPrice() >= minimumUSD,
             "Your donation is less than minimum donation amount"
         );*/
-        console.log(
-            "your donation value in USD : %s",
-            msg.value.getConversionRate(priceFeed)
-        );
-        require(
-            msg.value.getConversionRate(priceFeed) > 50,
-            "Your donation is less than minimum donation amount"
-        );
-        s_transactions.push(
-            Transaction(msg.sender, msg.value, block.timestamp)
-        );
+        console.log("your donation value in USD : %s", msg.value.getConversionRate(priceFeed));
+        require(msg.value.getConversionRate(priceFeed) > 50, "Your donation is less than minimum donation amount");
+        s_transactions.push(Transaction(msg.sender, msg.value, block.timestamp));
         s_bank += msg.value;
     }
 
     function cheapWithdraw() external onlyOwner {
         uint256 fundersLength = s_funders.length;
-        for (uint i = 0; i < fundersLength; i++) {
+        for (uint256 i = 0; i < fundersLength; i++) {
             s_addressToAmountFunded[s_funders[i]] = 0;
         }
         s_funders = new address[](0);
         delete s_transactions;
         address payable walletAddress = payable(msg.sender);
-        console.log(
-            "s_bank: %s, actual balance: %s",
-            s_bank,
-            address(this).balance
-        );
-        (bool success, ) = walletAddress.call{value: s_bank}("");
+        console.log("s_bank: %s, actual balance: %s", s_bank, address(this).balance);
+        (bool success,) = walletAddress.call{value: s_bank}("");
         require(success, "Failed to withdraw");
         s_bank = 0;
     }
 
     function withdraw() external onlyOwner {
-        for (uint i = 0; i < s_funders.length; i++) {
+        for (uint256 i = 0; i < s_funders.length; i++) {
             s_addressToAmountFunded[s_funders[i]] = 0;
         }
         s_funders = new address[](0);
         delete s_transactions;
         address payable walletAddress = payable(msg.sender);
-        console.log(
-            "s_bank: %s, actual balance: %s",
-            s_bank,
-            address(this).balance
-        );
-        (bool success, ) = walletAddress.call{value: s_bank}("");
+        console.log("s_bank: %s, actual balance: %s", s_bank, address(this).balance);
+        (bool success,) = walletAddress.call{value: s_bank}("");
         require(success, "Failed to withdraw");
         s_bank = 0;
     }
@@ -121,10 +103,8 @@ contract FundMe {
         return eth.getConversionRate(priceFeed);
     }
 
-    function getTxInfoByAddress(
-        address _address
-    ) external view returns (uint256 amount, uint256 time) {
-        for (uint i = 0; i < s_transactions.length; i++) {
+    function getTxInfoByAddress(address _address) external view returns (uint256 amount, uint256 time) {
+        for (uint256 i = 0; i < s_transactions.length; i++) {
             if (s_transactions[i].sender == _address) {
                 return (s_transactions[i].amount, s_transactions[i].time);
             }
